+++ Kotiln/payment-service/src/main/kotlin/core/paymentservice/repository/PaymentEventRepository.kt
@@ -3,4 +3,4 @@ package core.paymentservice.repository
 import core.paymentservice.domain.PaymentEvent
 import org.springframework.data.repository.reactive.ReactiveCrudRepository
 
-interface PaymentEventRepository : ReactiveCrudRepository<PaymentEvent, Int>
+interface PaymentEventRepository : ReactiveCrudRepository<PaymentEvent, Long>
