+++ Kotiln/payment-service/src/main/kotlin/core/paymentservice/repository/PaymentOrderRepository.kt
@@ -4,6 +4,6 @@ import core.paymentservice.domain.PaymentOrder
 import org.springframework.data.repository.reactive.ReactiveCrudRepository
 import reactor.core.publisher.Flux
 
-interface PaymentOrderRepository : ReactiveCrudRepository<PaymentOrder, Int> {
-    fun findByPaymentEventId(paymentEventId: Int): Flux<PaymentOrder>
+interface PaymentOrderRepository : ReactiveCrudRepository<PaymentOrder, Long> {
+    fun findByPaymentEventId(paymentEventId: Long): Flux<PaymentOrder>
 }
