+++ Kotiln/payment-service/src/main/kotlin/core/paymentservice/repository/ProductRepository.kt
@@ -3,4 +3,4 @@ package core.paymentservice.repository
 import core.paymentservice.domain.Product
 import org.springframework.data.repository.reactive.ReactiveCrudRepository
 
-interface ProductRepository : ReactiveCrudRepository<Product, Int>
+interface ProductRepository : ReactiveCrudRepository<Product, Long>
